fix float32 argument into flash_attn causing errors

+++ patch_flash_attn.py
@@ -83,7 +83,7 @@ def forward(
         x_unpad, indices, cu_q_lens, max_s = unpad_input(x, key_padding_mask)
         x_unpad = rearrange(
             x_unpad, "nnz (three h d) -> nnz three h d", three=3, h=nheads
-        )
+        ).type(torch.bfloat16)
         output_unpad = flash_attn_varlen_qkvpacked_func(
             x_unpad, cu_q_lens, max_s, 0.0, softmax_scale=None, causal=True
         )
